refactor: :building_construction: Separate my utils from Shadcn Svelte utils
