+++ src/lib/utils.ts
@@ -285,22 +285,3 @@ export type PrimitiveTableSectionAttributes = WithElementRef<
   HTMLAttributes<HTMLTableSectionElement>
 >;
 export type PrimitiveImgAttributes = WithElementRef<HTMLImgAttributes>;
-
-
-/**
-   * Convert a blob to array of bytes
-   */
-export async function blobToBytes(blob: Blob): Promise<Uint8Array> {
-  if (blob.bytes) return blob.bytes();
-  // Fallback to making bytes from array buffer
-  return new Uint8Array(await blob.arrayBuffer());
-}
-
-/**
-   * Convert a blob to array of bytes with a given type
-   */
-export async function blobChunksToBytes(chunks: Blob[], type = "audio/wav"): Promise<Uint8Array> {
-  const blob = chunks.length === 1 ? chunks[0]! : new Blob(chunks, { type: type });
-  return blobToBytes(blob);
-
-}
